+++ testkit/jvm/src/main/scala/busymachines/pureharm/testkit/util/PureharmTestPlatformSpecific.scala
@@ -14,9 +14,14 @@
  * limitations under the License.
  */
 
-package busymachines.pureharm
+package busymachines.pureharm.testkit.util
 
-/** @author Lorand Szakacs, https://github.com/lorandszakacs
-  * @since 26 Jun 2020
-  */
-package object testkit extends PureharmTestkitAliases
+import busymachines.pureharm.effects.pools._
+
+trait PureharmTestPlatformSpecific {
+  protected def defaultExecutionContext: ExecutionContextCT = _ecCT
+  protected def defaultFT:               ExecutionContextFT = _ecFT
+
+  private lazy val _ecCT = UnsafePools.cached("phtest-ec")
+  private lazy val _ecFT = UnsafePools.fixed("phtest-ft", 8)
+}